requirements.txt
numpy==1.22.2
pandas==1.4.2
Pillow==9.1.1
pydeck==0.7.1
requests==2.26.0
SPARQLWrapper==2.0.0
streamlit==1.5.0
